Add SMACSS-like shortcut syntax
In the SMACSS book, the authors recommend us to use the hierarcy
structure of nested blocks, and they recommend to call these blocks
using base name of the block as the prefix, to avoid names clashing and
messing up with the styles. This commit adds nifty shortcuts for that.
Have a look at the examples, they describe this better than thousands of
words.

Examples:

1.

  .header
    .+login
      label.-label Label
      input.-input= text_field_tag(...)

  is the same as:

  .header
    .login
      label.login_label Label
      input.login_input= text_field_tag(...)

2.

  .header
    .+navigation
      .+-main-menu
        ul.+-items
          li.+-item

  is the same as:

  .header
    .navigation
      .navigation_main-menu
        ul.navigation_main-menu_items
          li.navigation_main-menu_items_item

So it adds 3 special symbols-shortcuts: '.-', '.+' and '.+-'.
* By '.+' you specify the prefix, which will be appended to all the nested
  items, which have '.-'.
* By '.+-' you specify the new prefix, but create
  it appending old prefix to the current one.
* By '.-' you just appends the currently specified prefix to the
  provided class attribute